Move texture 2d/3d pin to their own files
Texture 3d accept size references
